Fix: Make Unversioned-DLL Target Build In Windows with Meson 1.7.0
As seen in github workflow, when we move from meson v1.6.1 to meson
v1.7.0, the custom target unversioned_dll, defined in
tests/meson.build and needed for test execution, no longer gets
built when windows.yaml invokes "ninja -C build".

As pointed out by @eli-schwartz (thank you), the meson 1.7.0 release
notes specify that test prerequesites will no longer be built by
default during the build phase (unversioned_dll is a test prerequisite
in this sense). They would be built in the test phase, however we have
specified --no-rebuild with "meson test", preventing unversioned_dll
being built in the test phase either.
cf. https://mesonbuild.com/Release-notes-for-1-7-0.html#test-targets-no-longer-built-by-default

In windows.yaml we now invoke build with:
    ninja -C build all meson-test-prereq
The "meson-test-prereq" causes unversioned_dll to be built in the
build phase.

Prior to this change -

Failing with meson 1.7.0:
https://github.com/pistacheio/pistache/actions/runs/13157241901/job/36735628578

Succeeding with meson v1.6.1:
https://github.com/pistacheio/pistache/actions/runs/12913131184/job/36009642858

ALSO - when unversioned_dll is built, we now copy to the unversioned
DLL name whenever the pistache DLL is newer than the unversioned DLL;
previously would only copy when the unversioned DLL did not exist,
which of course risks leaving a stale version of the unversioned DLL
in place if the pistache DLL has been rebuilt.

+++ src/server/listener.cc
@@ -731,8 +731,9 @@ namespace Pistache::Tcp
 
 #ifdef _IS_WINDOWS
 
-                unsigned long int timeout_in_ms = static_cast<unsigned int>(
-                    std::chrono::duration_cast<
+                unsigned long int timeout_in_ms =
+                    static_cast<unsigned long int>(
+                        std::chrono::duration_cast<
                         std::chrono::milliseconds>(sslHandshakeTimeout_)
                         .count());
